Implements #11

+++ recpack/helpers.py
@@ -18,7 +18,7 @@ def create_csr_matrix_from_pandas_df(df, item_ix, user_ix):
     values = np.ones(num_entries)
     indices = zip(*df.loc[:, [row_ix, column_ix]].values)
     shape = df[row_ix].max() + 1, df[column_ix].max() + 1
-    sparse_matrix = scipy.sparse.csr_matrix((values, indices), shape=shape, dtype=np.float32)
+    sparse_matrix = scipy.sparse.csr_matrix((values, indices), shape=shape, dtype=np.int32)
     sparsity = sparse_matrix.getnnz() / (sparse_matrix.shape[0] * sparse_matrix.shape[1]) * 100
     logger.info(f"Sparsity is {sparsity}")
     return sparse_matrix
